Use sub-rule as separate rule.
