+++ datafusion/sqllogictest/test_files/groupby.slt
@@ -1962,7 +1962,7 @@ SortPreservingMergeExec: [col0@0 ASC NULLS LAST]
 ------SortExec: expr=[col0@0 ASC NULLS LAST]
 --------CoalesceBatchesExec: target_batch_size=8192
 ----------RepartitionExec: partitioning=Hash([col0@0, col1@1, col2@2], 4), input_partitions=4
-------------AggregateExec: mode=Partial, gby=[col0@0 as col0, col1@1 as col1, col2@2 as col2], aggr=[FIRST_VALUE(r.col1)], ordering_mode=PartiallyOrdered
+------------AggregateExec: mode=Partial, gby=[col0@0 as col0, col1@1 as col1, col2@2 as col2], aggr=[LAST_VALUE(r.col1)], ordering_mode=PartiallyOrdered
 --------------SortExec: expr=[col0@3 ASC NULLS LAST]
 ----------------CoalesceBatchesExec: target_batch_size=8192
 ------------------HashJoinExec: mode=Partitioned, join_type=Inner, on=[(col0@0, col0@0)]
